requirements.txt
imagehash
pillow
streamlink
flask
redis